requirements.txt
alembic==1.15.2
bleach==6.2.0
blinker==1.9.0
boto3==1.39.7
botocore==1.39.7
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.1.8
coverage==7.10.0
Deprecated==1.2.18
dnspython==2.7.0
email_validator==2.2.0
factory_boy==3.3.3
Faker==37.4.2
Flask==3.1.0
Flask-Limiter==3.12
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-Moment==1.0.6
flask-paginate==2024.4.12
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.1.1
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jinja2_markdown==0.0.3
jmespath==1.0.1
limits==5.4.0
Mako==1.3.9
Markdown==3.8
markdown-it-py==3.0.0
markdown2==2.5.3
MarkupSafe==3.0.2
mdurl==0.1.2
ordered-set==4.1.0
packaging==24.2
pluggy==1.6.0
psycopg2-binary==2.9.10
Pygments==2.19.2
pytest==8.4.1
pytest-cov==6.2.1
pytest-flask==1.3.0
pytest-flask-sqlalchemy==1.1.0
pytest-mock==3.14.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
PyYAML==6.0.2
requests==2.32.4
responses==0.25.7
rich==13.9.4
s3transfer==0.13.0
six==1.17.0
SQLAlchemy==2.0.40
typing_extensions==4.13.0
tzdata==2025.2
urllib3==2.5.0
webencodings==0.5.1
Werkzeug==3.1.3
wrapt==1.17.2
WTForms==3.2.1
